added Menu functionality in separate component
